pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset-index-cli"
version = "0.1.0"
description = "Atomic CLI for frontmatter-based asset index management"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [
    {name = "hezi-ywt"},
]
keywords = ["cli", "asset", "index", "frontmatter", "markdown", "yaml", "content-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/hezi-ywt/asset-index"
Repository = "https://github.com/hezi-ywt/asset-index"
Issues = "https://github.com/hezi-ywt/asset-index/issues"

[project.scripts]
asset-index = "asset_index.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
asset_index = []
